+++ src/app/main/apps/tarefas/tarefas.component.ts
@@ -752,6 +752,10 @@ export class TarefasComponent implements OnInit, OnDestroy, AfterViewInit {
     // @ Public methods
     // -----------------------------------------------------------------------------------------------------
 
+    limpaFiltros(): void {
+        this.buscarTodas = false;
+    }
+
     reload(params): void {
         this.novaTarefa = false;
         const nparams = {
@@ -779,9 +783,7 @@ export class TarefasComponent implements OnInit, OnDestroy, AfterViewInit {
                 modulo: generoParam,
                 mostrarApagadas: true
             };
-        }
-
-        if (this.typeHandle === 'minhas-tarefas') {
+        } else if (this.typeHandle === 'minhas-tarefas') {
             nparams.filter = {
                 'usuarioResponsavel.id': 'eq:' + this._profile.id,
                 'dataHoraConclusaoPrazo': 'isNull'
@@ -822,57 +824,6 @@ export class TarefasComponent implements OnInit, OnDestroy, AfterViewInit {
                     mostrarApagadas: true
                 };
             }
-        }
-
-        if (this.typeHandle === 'compartilhadas') {
-
-            if (this.routerState.params['targetHandle'] === 'outros-usuarios') {
-                nparams.filter = {
-                    'compartilhamentos.usuario.id': 'eq:' + this._profile.id,
-                    'dataHoraConclusaoPrazo': 'isNull'
-                };
-            }
-
-            if (this.routerState.params['targetHandle'] === 'meus-compartilhamentos') {
-                nparams.filter = {
-                    'compartilhamentos.tarefa.usuarioResponsavel.id': 'eq:' + this._profile.id,
-                    'dataHoraConclusaoPrazo': 'isNull'
-                };
-            }
-        }
-
-        if (this.typeHandle === 'concluidas') {
-            nparams.filter = {
-                'usuarioResponsavel.id': 'eq:' + this._profile.id,
-                'dataHoraConclusaoPrazo': 'isNotNull'
-            };
-        }
-
-        if (this.typeHandle === 'enviadas') {
-            nparams.filter = {
-                'criadoPor.id': 'eq:' + this._profile.id,
-                'usuarioResponsavel.id': 'neq:' + this._profile.id,
-            };
-        }
-
-        if (this.typeHandle === 'coordenacao') {
-            nparams.filter = {
-                dataHoraConclusaoPrazo: 'isNull'
-            };
-            const routeTargetParam = of('targetHandle');
-            routeTargetParam.subscribe((targetParam) => {
-                nparams.filter['setorResponsavel.id'] = `eq:${this.routerState.params[targetParam]}`;
-            });
-        }
-
-        if (this.typeHandle === 'assessor') {
-            nparams.filter = {
-                dataHoraConclusaoPrazo: 'isNull'
-            };
-            const routeTargetParam = of('targetHandle');
-            routeTargetParam.subscribe((targetParam) => {
-                nparams.filter['usuarioResponsavel.id'] = `eq:${this.routerState.params[targetParam]}`;
-            });
         }
 
         nparams['filter'] = {