+++ CHANGELOG.md
@@ -1072,3 +1072,4 @@
 * Correção em bug ao trocar a pasta enquanto tarefas se encontram filtradas
 * Correção em bug onde mais de uma requisição de assinatura era enviada ao backend, resultando em duplicação de assinaturas de minuta
 * Correção em outro local onde a assinatura poderia ocorrer de forma duplicada
+* Correção de erros na lista de tarefas após implementação dos filtros de buscar todas